Add completed README. Remove Develop folder to simplify installation process for user.
