--- Develop/models/Category.js
@@ -1,31 +1,0 @@
-// Require Model and DataTypes from sequelize library.
-const { Model, DataTypes } = require('sequelize');
-
-// Require database connection from config.js.
-const sequelize = require('../config/connection.js');
-
-class Category extends Model {}
-
-Category.init(
-  {
-    id: {
-      type: DataTypes.INTEGER, 
-      allowNull: false,
-      primaryKey: true,
-      autoIncrement: true
-    }, 
-    category_name: {
-      type: DataTypes.STRING,
-      allowNull: false
-    }
-  },
-  {
-    sequelize,
-    timestamps: false,
-    freezeTableName: true,
-    underscored: true,
-    modelName: 'category',
-  }
-);
-
-module.exports = Category;
