--- Develop/models/Tag.js
@@ -1,30 +1,0 @@
-// Require Model and DataTypes from sequelize library.
-const { Model, DataTypes } = require('sequelize');
-
-// Require database connection from config.js.
-const sequelize = require('../config/connection.js');
-
-class Tag extends Model {}
-
-Tag.init(
-  {
-    id: {
-      type: DataTypes.INTEGER, 
-      allowNull: false, 
-      primaryKey: true, 
-      autoIncrement: true
-    }, 
-    tag_name: {
-      type: DataTypes.STRING
-    }
-  },
-  {
-    sequelize,
-    timestamps: false,
-    freezeTableName: true,
-    underscored: true,
-    modelName: 'tag',
-  }
-);
-
-module.exports = Tag;
