+++ models/Product.js
@@ -40,7 +40,7 @@ Product.init(
             references: {
                 model: 'category',
                 key: 'id',
-            }, 
+            },
         }
     },
     {
