+++ routes/api/tag-routes.js
@@ -77,7 +77,7 @@ router.delete('/:id', async (req, res) => {
         }
 
         return res.status(200).json(`Number of rows deleted: ${JSON.stringify(tagData)}`);
-        
+
     } catch (err) {
         res.status(500).json(err);
     }
