+++ server.js
@@ -13,6 +13,6 @@ app.use(routes);
 
 // Sync sequelize models to the database, then turn on the server.
 sequelize.sync().then(() => {
-  app.listen(PORT, () => console.log(`App listening on port ${PORT}`));
+    app.listen(PORT, () => console.log(`App listening on port ${PORT}`));
 });
